Rename Tree to TreeMenu
